Former-commit-id: 756240fbb1fe492d581739420c464970b317ca1f

+++ toolboxes/registration/optical_flow/multiresRegistrationSolver.h
@@ -228,9 +228,9 @@ namespace Gadgetron{
 
       for( unsigned int i=0; i<num_multires_levels_; i++ ){
 
-        dims /= 2UL;
+        dims /= (size_t)2;
 
-        if( weak_less( dims, 12UL*ones ) ){
+        if( weak_less( dims, (size_t)12*ones ) ){
           throw std::runtime_error("multiresRegistrationSolver::padding_required : resolution too low. Too many multiresolution levels specified?");
         }
